+++ centrifuge-js/src/modules/metadata.ts
@@ -1,112 +1,5 @@
 import { first, from, Observable } from 'rxjs'
 import Centrifuge from '..'
-interface TrancheFormValues {
-  tokenName: string
-  symbolName: string
-  interestRate: number | ''
-  minRiskBuffer: number | ''
-  minInvestment: number | ''
-}
-interface RiskGroupFormValues {
-  groupName: string
-  advanceRate: number | ''
-  fee: number | ''
-  probabilityOfDefault: number | ''
-  lossGivenDefault: number | ''
-  discountRate: number | ''
-}
-interface WriteOffGroupFormValues {
-  days: number | ''
-  writeOff: number | ''
-}
-
-export interface PoolMetadataInput {
-  // details
-  poolIcon: string | null
-  poolName: string
-  assetClass: string
-  currency: string
-  maxReserve: number | ''
-  epochHours: number | ''
-  epochMinutes: number | ''
-
-  // issuer
-  issuerName: string
-  issuerLogo: string | null
-  issuerDescription: string
-
-  executiveSummary: string | null
-  website: string
-  forum: string
-  email: string
-
-  // tranche
-  tranches: TrancheFormValues[]
-  riskGroups: RiskGroupFormValues[]
-  writeOffGroups: WriteOffGroupFormValues[]
-}
-
-export type PoolStatus = 'open' | 'upcoming' | 'hidden'
-export type PoolCountry = 'us' | 'non-us'
-export type NonSolicitationNotice = 'all' | 'non-us' | 'none'
-export type PoolMetadata = {
-  pool: {
-    name: string
-    icon: string
-    asset: {
-      class: string
-    }
-    issuer: {
-      name: string
-      description: string
-      email: string
-      logo: string
-    }
-    links: {
-      executiveSummary: string
-      forum: string
-      website: string
-    }
-    status: PoolStatus
-  }
-  tranches: Record<
-    string,
-    {
-      name: string
-      symbol: string
-      minInitialInvestment: string
-    }
-  >
-  riskGroups: {
-    name: string | undefined
-    advanceRate: string
-    interestRatePerSec: string
-    probabilityOfDefault: string
-    lossGivenDefault: string
-    discountRate: string
-  }[]
-  // Not yet implemented
-  // onboarding: {
-  //   live: boolean
-  //   agreements: {
-  //     name: string
-  //     provider: 'docusign'
-  //     providerTemplateId: string
-  //     tranche: string
-  //     country: 'us | non-us'
-  //   }[]
-  //   issuer: {
-  //     name: string
-  //     email: string
-  //     restrictedCountryCodes: string[]
-  //     minInvestmentCurrency: number
-  //     nonSolicitationNotice: 'all' | 'non-us' | 'none'
-  //   }
-  // }
-  // bot: {
-  //   channelId: string
-  // }
-}
 
 export function getMetadataModule(inst: Centrifuge) {
   function getMetadata<T = any>(uri: string): Observable<T | T[] | null> {
@@ -117,7 +10,7 @@ export function getMetadataModule(inst: Centrifuge) {
     return inst.getMetadataObservable<T>(url)
   }
 
-  function pinFile(metadata: any): Observable<{ uri: string; ipfsHash: string }> {
+  function pinFile(metadata: Record<any, any>): Observable<{ uri: string; ipfsHash: string }> {
     if (!inst.config?.pinFile) {
       console.error('pinFile must be set in config to use this feature')
       return from([])
@@ -131,7 +24,7 @@ export function getMetadataModule(inst: Centrifuge) {
     ).pipe(first())
   }
 
-  function pinJson(metadata: any): Observable<{ uri: string; ipfsHash: string }> {
+  function pinJson(metadata: Record<any, any>): Observable<{ uri: string; ipfsHash: string }> {
     if (!inst.config.pinJson) {
       console.error('pinJson must be set in config to use this feature')
       return from([])
